fix: moved file structure, code in src, non-code outside
