+++ server/src/routes/MealSchedule.js
@@ -1,7 +1,7 @@
 const express = require('express');
 
 const router = express.Router();
-const { schedule } = require('./mock/MealScheduleMockData');
+const { schedule } = require('../mock/MealScheduleMockData');
 
 const schedules = [
   {
